+++ src/esm/drawutilssvg.d.ts
@@ -48,7 +48,8 @@
  * @modified 2024-03-10 Fixing some types for Typescript 5 compatibility.
  * @modified 2024-07-24 Caching custom style defs in a private buffer variable.
  * @modified 2025-11-14 Fixing a bug in the CSS `mix-blend-mode` property handling (caused a runtime error).
- * @version  1.6.11
+ * @modified 2026-01-04 Adding `lineJoin` attribute to the methods' `StrokeOptions` param.
+ * @version  1.7.0
  **/
 import { Polygon } from "./Polygon";
 import { Vertex } from "./Vertex";
@@ -747,11 +748,12 @@ export declare class drawutilssvg implements DrawLib<void | SVGElement> {
      * @param {Polygon} polygon - The polygon to draw.
      * @param {string} color - The CSS color to draw the polygon with.
      * @param {number=} lineWidth - (optional) The line width to use; default is 1.
+     * @param {StrokeOptions=} strokeOptions - (optional) Stroke settings to use.
      * @return {void}
      * @instance
      * @memberof drawutilssvg
      */
-    polygon(polygon: Polygon, color: string, lineWidth?: number): SVGElement;
+    polygon(polygon: Polygon, color: string, lineWidth?: number, strokeOptions?: StrokeOptions): SVGElement;
     /**
      * Draw a polygon line (alternative function to the polygon).
      *
@@ -876,6 +878,7 @@ export declare class drawutilssvg implements DrawLib<void | SVGElement> {
      */
     static transformPathData(data: SVGPathParams, offset: XYCoords, scale: XYCoords): void;
     private static nodeSupportsLineDash;
+    private static nodeSupportsLineJoin;
     /**
      * Creates a basic <line> node with start and end coordinates. The created node will not
      * be bound to any root node.